move block index logic to a ContentToClaims mapper
